Applied frame based titan spawning as per 151 on Waves, Endless and KillTitans gamemode. Removed TITAN_CONTROLLER
