+++ Assembly-CSharp.csproj
@@ -136,6 +136,7 @@
      <Compile Include="Assets\Scripts\Characters\GROUP.cs" />
      <Compile Include="Assets\Scripts\Characters\HitBox.cs" />
      <Compile Include="Assets\Scripts\Characters\Horse\Horse.cs" />
+     <Compile Include="Assets\Scripts\Characters\Horse\HorseController.cs" />
      <Compile Include="Assets\Scripts\Characters\Human\BODY_PARTS.cs" />
      <Compile Include="Assets\Scripts\Characters\Human\BUFF.cs" />
      <Compile Include="Assets\Scripts\Characters\Human\Bullet.cs" />
@@ -200,7 +201,6 @@
      <Compile Include="Assets\Scripts\Characters\Titan\PlayerTitan.cs" />
      <Compile Include="Assets\Scripts\Characters\Titan\RockScript.cs" />
      <Compile Include="Assets\Scripts\Characters\Titan\RockThrow.cs" />
-     <Compile Include="Assets\Scripts\Characters\Titan\TITAN_CONTROLLER.cs" />
      <Compile Include="Assets\Scripts\Characters\Titan\TITAN_EREN.cs" />
      <Compile Include="Assets\Scripts\Characters\Titan\TITAN_SETUP.cs" />
      <Compile Include="Assets\Scripts\Characters\Titan\TitanBody.cs" />
