+++ Assets/Resources/horse.prefab
@@ -1888,31 +1888,14 @@ MonoBehaviour:
   m_GameObject: {fileID: 1965692679872925}
   m_Enabled: 1
   m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1a11a4234755e5c4f9cc345b44d9961f, type: 3}
+  m_Script: {fileID: 11500000, guid: 4f4174f7a6503194ca5c13adc51d7579, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  bite: 0
-  bitel: 0
-  biter: 0
-  chopl: 0
-  chopr: 0
-  choptl: 0
-  choptr: 0
-  cover: 0
   currentCamera: {fileID: 0}
   currentDirection: 0
-  grabbackl: 0
-  grabbackr: 0
-  grabfrontl: 0
-  grabfrontr: 0
-  grabnapel: 0
-  grabnaper: 0
   inputManager: {fileID: 0}
   isAttackDown: 0
   isAttackIIDown: 0
-  isHorse: 1
-  isJumpDown: 0
-  isSuicide: 0
   isWALKDown: 0
   sit: 0
   targetDirection: 0
