+++ Assets/Scripts/Characters/Horse/HorseController.cs.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 1a11a4234755e5c4f9cc345b44d9961f
-timeCreated: 1582403858
-licenseType: Free
+guid: 4f4174f7a6503194ca5c13adc51d7579
 MonoImporter:
+  externalObjects: {}
   serializedVersion: 2
   defaultReferences: []
   executionOrder: 0
